feat: allow JSON in `BOXEDNODE_{CONFIGURE,MAKE}_ARGS`
In mongosh, we introduced `NODE_JS_CONFIGURE_ARGS` (not
remembering that boxednode already provides this capability).
We should just use a single consistent env var throughout
the build process.

This adds JSON array support for these environment variables
for cases in which the list items themselves contain commas,
and, as a drive-by fix, avoids mutating the input options
argument.