chore: delete precompiled headers on windows and node24>
Compiling with a snapshot requires compiling twice: a base image
and the image with the snapshot embedded. In that situation, clang-cl
complains that there are precompiled headers that changed between
compilations and does not refresh them, but kills the compilation
process with an error. Due to this, before attempting the second
compilation, we will delete all pch files.