chore: Generate the GHA workflow from a template
Because we can keep adding more tests, maintaining the GHA workflow
can become a bit cumbersome. This script generates the GHA file
from a template, and adds all the test cases in a matrix. We do this
because each test is expensive in time and disk space, so we do
sometimes hit the GHA limit of 6h or the we fill the disk.