chore: ensure that any output directory is cleaned up before compiling
Right now it's failing on the first compilation of the second test, as
we are not doing any clean up. To ensure that we are in a consistent
state in Windows, before attempting compiling, we will make sure we
don't have any old compilation artifact.

This might slow down Windows compilations as they can't reuse cached
files.